requirements.txt
httpx>=0.27.0
woocommerce>=3.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.0.0
